Introduce the tile module by moving modules from ::io to ::tile
