+++ maplibre-demo/src/main.rs
@@ -2,24 +2,22 @@ use maplibre::benchmarking::tessellation::{IndexDataType, OverAlignedVertexBuffe
 use maplibre::coords::{WorldTileCoords, ZoomLevel};
 use maplibre::error::Error;
 use maplibre::headless::HeadlessMapWindowConfig;
-use maplibre::io::pipeline::Processable;
-use maplibre::io::pipeline::{PipelineContext, PipelineProcessor};
 
 use maplibre::io::source_client::{HttpClient, HttpSourceClient};
-use maplibre::io::tile_pipelines::build_vector_tile_pipeline;
-use maplibre::io::tile_repository::StoredLayer;
-use maplibre::io::{RawLayer, TileRequest};
+use maplibre::io::TileRequest;
 
+use maplibre::headless::utils::HeadlessPipelineProcessor;
 use maplibre::platform::http_client::ReqwestHttpClient;
 use maplibre::platform::run_multithreaded;
 use maplibre::platform::schedule_method::TokioScheduleMethod;
 use maplibre::render::settings::{RendererSettings, TextureFormat};
 use maplibre::render::ShaderVertex;
+use maplibre::tile::pipeline::PipelineContext;
+use maplibre::tile::pipeline::Processable;
+use maplibre::tile::tile_pipelines::build_vector_tile_pipeline;
 use maplibre::window::{EventLoop, WindowSize};
 use maplibre::MapBuilder;
 use maplibre_winit::winit::WinitMapWindowConfig;
-
-use maplibre::headless::utils::HeadlessPipelineProcessor;
 use std::collections::HashSet;
 
 #[cfg(feature = "trace")]
